fix(groups): forces content accessibility to members_only for invisible groups
When user selects "Group members only" for group visibility, javascript sets
the group content accessibility field to "Members only" and disables it.

For existing groups the above is done also on server side.

When editing an existing public group, a warning is added under the group
content accessibility input saying that changing the setting won't change
the access permission of existing group content.