fix(plugins): discussion replies no longer missing/reversed order on river
A switch to elgg_list_entities() caused the URL offset param to influence
the display of inline replies, causing the river not to show the latest on
later pages. Here we move back to ege() but we add a warning in the docs for
ele() about this behavior. This also reverses the order of the replies so it
matches the river comments.

Fixes #7801, #7668